move duplicate cert test to main_test.py
